feat: set environment variables for hooks
Provide the package root, source files and the target directory as
environment variables to the package hooks. This is makes hooks
that run prior to installation or after uninstallation when no package
state exists more useful, as the package state path was previously the
only value provided to the hook (as an argument).